move RotateStrategy to Rotate
